Reworked the HUD

+++ src/main/java/rsmm/fabric/client/gui/element/TextElement.java
@@ -180,7 +180,7 @@ public class TextElement implements IElement {
 	}
 	
 	@Override
-	public List<Text> getTooltip(double mouseX, double mouseY) {
+	public List<Text> getTooltip(int mouseX, int mouseY) {
 		return tooltipSupplier.get();
 	}
