+++ src/main/java/rsmm/fabric/client/gui/widget/Button.java
@@ -205,7 +205,7 @@ public class Button extends ButtonWidget implements IElement, IButton {
 	}
 	
 	@Override
-	public List<Text> getTooltip(double mouseX, double mouseY) {
+	public List<Text> getTooltip(int mouseX, int mouseY) {
 		return tooltipSupplier.get();
 	}
