+++ src/main/java/rsmm/fabric/client/gui/widget/Slider.java
@@ -153,7 +153,7 @@ public class Slider extends SliderWidget implements IButton {
 	}
 	
 	@Override
-	public List<Text> getTooltip(double mouseX, double mouseY) {
+	public List<Text> getTooltip(int mouseX, int mouseY) {
 		return tooltipSupplier.get();
 	}
